javascript combined
