+++ core/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "prestart": "",
     "start": "goapp serve",
-    "pretest" : "goapp get github.com/stretchr/testify/assert",
+    "pretest": "goapp get github.com/stretchr/testify/assert",
     "test": "goapp test Xtern-Matching/tests",
     "test-front": "karma start karma.conf.js"
   },
@@ -31,12 +31,12 @@
     "angular-legacy-sortablejs": "^0.4.1",
     "angular-sanitize": "^1.5.8",
     "angular-ui-router": "^0.2.18",
+    "browserify": "^14.0.0",
     "core-js": "^2.3.0",
     "gulp": "^3.9.1",
     "http": "0.0.0",
-    "jquery": "^2.2.1",
     "ng-sortable": "^1.3.6",
-	  "pdfjs-dist" : "^1.5.500"
+    "pdfjs-dist": "^1.5.500"
   },
   "devDependencies": {
     "angular-mocks": "^1.5.0",
