+++ core/public/js/index.js
@@ -1,0 +1,7 @@
+'use strict';
+
+//require('./login.js');
+//require('./helpers.js');
+//require('./services.js');
+//require('./directives.js');
+//require('./temp_globals.js');
