+++ core/public/js/services.js
@@ -1,5 +1,4 @@
 'use strict';
-(function () {
     var app = angular.module('DataManager', []);
     var host = "http://localhost:8080/";
     app
@@ -518,4 +517,4 @@
                 });
             };
         }]);
-})();
+
