+++ core/public/modules/account_control/index.js
@@ -1,0 +1,6 @@
+/**
+ * Created by Steven on 2/13/2017.
+ */
+'use strict';
+
+require('./techpoint.accountControl.controller.js');
