+++ core/public/modules/dashboard/index.js
@@ -1,0 +1,7 @@
+'use strict';
+
+require('./dashboard.controller.js');
+require('./dashboard.directives.js');
+require('./company.dashboard.controller.js');
+require('./reviewer.dashboard.controller.js');
+require('./techpoint.dashboard.controller.js');
