+++ core/public/modules/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+
+require('./account_control/');
+require('./dashboard/');
+require('./student_profile/');
+require('./login/');
+require('./recruting/');
+require('./reviewer_controls/');
