+++ core/public/modules/login/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+//require('./login.js');
+require('./techpoint.login.controller.js');
+require('./company.login.controller.js');
+require('./reviewer.login.controller.js');
