+++ core/public/modules/recruting/index.js
@@ -1,0 +1,3 @@
+'use strict';
+
+require('./company.recruting.js');
