+++ core/public/modules/reviewer_controls/index.js
@@ -1,0 +1,3 @@
+'use strict';
+
+require('./techpoint.reviewerControls.controller.js');
