+++ core/public/modules/student_profile/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+require('./studentDataPage.controller.js');
+require('./studentProfile.directives.js');
+require('./comments/comment.controller.js');
+require('./tech_labels/techLabels.controller.js');
+require('./techpoint.studentProfile.controller.js');
+require('./company.studentProfile.controller.js');
+require('./reviewer.studentProfile.controller.js');
