+++ core/public/modules/student_profile/partials/studentProfile.html
@@ -148,11 +148,8 @@
 </div> -->
 <div class="vertical-spacer"></div>
 
-<script src="node_modules/pdfjs-dist/build/pdf.js">
 
-<script src="public/js/pdfobject.js"></script>
 <style>
     #the-canvas { height: 700px; }
     .pdfobject { border: 1px solid #666; }
 </style>
-<link rel="stylesheet" type="text/css" href="public/css/studentProfile.css">
